requirements.txt
efficientnet_pytorch==0.7.1
grad_cam==1.3.7
imutils==0.5.4
matplotlib==3.2.2
numpy==1.20.2
opencv_python==4.3.0.36
pandas==1.2.5
Pillow==9.3.0
pretrainedmodels==0.7.4
scikit_learn==1.1.3
SimpleITK==2.2.0
tensorboardX==2.5.1
torch==1.9.0
torchvision==0.10.0
